refactor(imports): remove remaining internal forwarding layers
Follow up on Chloe-JY’s re-export cleanup and PR #1350. Preserve public APIs and defer large state facades.
