--- src/engines/Simulator/exports.ts
@@ -1,10 +1,0 @@
-/**
- * ActivitySimulator Exports
- */
-export { default as ActivitySimulator } from "./ActivitySimulator";
-export { default as SimulatorContentArea } from "./SimulatorMainPane";
-export { default as ActivitySimulatorGrid } from "./ActivitySimulatorGrid";
-export * from "./config";
-
-// Hooks
-export { useSimEventRenderer } from "./hooks/useSimEventRenderer";
