--- src/modules/MainApp/Integrations/DevTools/playground/single-event/index.ts
@@ -1,1 +1,0 @@
-export { SingleEventPreview } from "./SingleEventPreview";
