--- src/modules/MainApp/Settings/subpages/BackgroundPage/components/index.ts
@@ -1,1 +1,0 @@
-export { ColorSection } from "./ColorSection";
