--- src/modules/MainApp/Settings/subpages/BackgroundPage/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useBackgroundSettings } from "./useBackgroundSettings";
