--- src/modules/WorkStation/CodeEditor/Panels/EditorMainPane/content/CodeViewerContent/hooks/index.ts
@@ -1,4 +1,0 @@
-/**
- * Hooks for CodeViewerContent
- */
-export { useCodeViewerHandlers } from "./useCodeViewerHandlers";
