--- src/scaffold/ContextMenu/variants/TextSelectionDropdown/exports.ts
@@ -1,8 +1,0 @@
-/**
- * TextSelectionDropdown Exports
- */
-
-export { default as TextSelectionDropdown } from "./index";
-export { useTextSelectionDropdown } from "./useTextSelectionDropdown";
-export * from "./config";
-export * from "./types";
