--- src/scaffold/ContextMenu/variants/index.ts
@@ -1,10 +1,0 @@
-/**
- * ContextMenu Variants - Barrel Exports
- *
- * Centralized exports for all context menu variants
- */
-
-export {
-  TextSelectionDropdown,
-  useTextSelectionDropdown,
-} from "./TextSelectionDropdown/exports";
