+++ src/scaffold/GlobalSpotlight/palettes/WorkspacePalette/index.tsx
@@ -16,6 +16,7 @@ import { useTranslation } from "react-i18next";
 import { repoApi } from "@src/api/tauri/repo";
 import Message from "@src/components/Message";
 import { HugeiconsIcon } from "@src/icons";
+import { useSelector as useSelectorKernel } from "@src/scaffold/GlobalSpotlight/hooks/selectors/useSelector";
 import { cachedReposAtom } from "@src/store/repo";
 import { addWorkspaceInitialStageAtom } from "@src/store/ui/overlayAtom";
 import { spotlightShowPathAtom } from "@src/store/ui/spotlightShowPathAtom";
@@ -43,7 +44,6 @@ import { PaletteBody, ShellFooterAction, SpotlightShell } from "../../shell";
 import type { RepoItem, SpotlightItem } from "../../types";
 import { AddWorkspaceModalShell } from "../AddWorkspaceModalShell";
 import { REPO_PALETTE_CONFIG } from "../config";
-import { useSelectorKernel } from "../core";
 import { buildOpenPathItem } from "./pathActionItem";
 import { importWorkspacePath } from "./pathImport";
 import { buildPinnedWorkspaceActions } from "./pinnedActions";
