--- src/scaffold/GlobalSpotlight/palettes/core/index.ts
@@ -1,7 +1,0 @@
-/**
- * Selector Core
- *
- * Shared kernel used by all palettes. Chrome (surface/portal/footer) is
- * provided by SpotlightShell in `../../shell`.
- */
-export { useSelector as useSelectorKernel } from "../../hooks/selectors/useSelector";
