--- src/scaffold/NavigationSidebar/contexts/index.ts
@@ -1,7 +1,0 @@
-/**
- * NavigationSidebar Contexts
- */
-export {
-  useForceVisibleSidebar,
-  ForceVisibleSidebarProvider,
-} from "./ForceVisibleContext";
