--- src/store/config/index.ts
@@ -1,6 +1,0 @@
-/**
- * App configuration atoms (IDE prefs, chat appearance, internal UI editor paths).
- * Distinct from `settings/` (JSONC file); may derive from settings where applicable.
- */
-
-export * from "./configAtom";
