--- src/store/platform/index.ts
@@ -1,6 +1,0 @@
-/**
- * Platform / developer tooling: dev mode toggle, system dependency scan cache.
- */
-
-export * from "./devModeAtom";
-export * from "./systemDepsAtom";
